Add backslash at the end of the contest URL.
Add backslash at the end of the contest URL when clicking the contest card.

This is done since there will be a backslash appended at the end of the contest
URL when clicking the 'Overview' button in the contest page. To ensure that
relative URL destination is consistent whether the contest overview is visited
before or after clicking any contest button, we also append a backslash at the
end of the contest URL before clicking any contest button.